requirements.txt
absl-py==0.11.0
aiohttp==3.6.3
asari==0.0.4
astor==0.8.1
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
dill==0.3.3
discord.py==1.5.1
future==0.18.2
gast==0.2.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
googletrans @ git+https://github.com/alainrouillon/py-googletrans@15b236f3b96fc9accbe427c0b39bbe1a05bc9338
grpcio==1.34.0
h11==0.9.0
h2==3.2.0
h5py==3.1.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==3.3.0
Janome==0.3.7
joblib==1.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lxml==4.6.2
Markdown==3.3.3
mojimoji==0.0.11
multidict==4.7.6
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.5
promise==2.3
protobuf==3.14.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.6
scikit-learn==0.22.2
scipy==1.5.4
six==1.15.0
sniffio==1.2.0
soupsieve==2.1
tensorboard==2.0.2
tensorflow==2.0.3
tensorflow-datasets==1.2.0
tensorflow-estimator==2.0.1
tensorflow-metadata==0.26.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.5.1
zipp==3.4.0